Deploy an "echoserver" testware
An echo server based on `nc` that runs in a pod that can be deployed alongside the chart under test.
It should allow us to send message lines to it and test how stdout logs from it will end up in elasticsearch.
The echo server is proxied via `kubectl port-forward` which will allow interaction with the pod/service in-cluster via
localhost tcp connection

Also:

- introduce a `wait_until` method for more reactive programming
-